split model_unix types and functions
